removed secret key

+++ dmgilmour.py
@@ -10,7 +10,6 @@ import bcrypt
 from getpass import getpass
 from appAlgoTime import Algo
 from flask import Flask, request, abort, url_for, redirect, session, render_template, flash, jsonify
-from flask_sqlalchemy import SQLAlchemy
 
 
 app = Flask(__name__)
@@ -18,8 +17,6 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///chat.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
 algo = Algo()
-
-master_secret_key = 'Senor Design'
 
 """
 class User(db.Model):
@@ -40,58 +37,6 @@ class Bike(db.Model):
         self.ip = ip
 """
 
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.column(db.String(40))
-    password = db.column(db.String(70))
-    salt = db.column(db.String(40))
-
-    def __init__(self, username, password, salt):
-        self.username = username
-        self.password = password
-        self.salt = salt
-
-
-class Location(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    lat = db.Column(db.String(50))
-    lon = db.Column(db.String(50))
-    time = db.Column(db.DateTime)
-    user_ip = db.Column(db.String(50))
-
-    def __init__(self, lat, lon, time, user_ip):
-        self.lat = lat
-        self.lon = lon
-        self.time = time
-        self.user_ip = user_ip
-
-@app.cli.command("initdb")
-def initdb_command():
-    db.drop_all()
-    db.create_all()
-
-    db.session.commit()
-
-    """
-    db.session.add(Location("lat1", "lat1", datetime.datetime.utcnow(), "ip1"))
-    time.sleep(.2)
-    db.session.add(Location("lat2", "lat2", datetime.datetime.utcnow(), "ip2"))
-    time.sleep(.2)
-    db.session.add(Location("lat3", "lat3", datetime.datetime.utcnow(), "ip3"))
-    time.sleep(.2)
-    db.session.add(Location("lat4", "lat4", datetime.datetime.utcnow(), "ip1"))
-    time.sleep(.2)
-    db.session.add(Location("lat5", "lat5", datetime.datetime.utcnow(), "ip5"))
-    time.sleep(.2)
-    db.session.add(Location("lat6", "lat6", datetime.datetime.utcnow(), "ip6"))
-
-    db.session.add(User("122.122.122.122"))
-
-    db.session.commit()
-
-    print(Location.query.filter_by(user_ip="ip1").order_by('-id').first().lat)
-"""
-
 
 
 @app.route("/login", methods = ["GET", "POST"])
@@ -108,7 +53,7 @@ def login():
             username = user[0]
             salt = user[2].encode('utf-8')
             print(salt)
-            combopass = (request.form["pass"] + salt + master_secret_key).encode('utf-8')
+            combopass = (request.form["pass"] + salt).encode('utf-8')
             password = bcrypt.hashpw(combopass, salt)
             if password == user[1]:
                 session["user"] = username
@@ -131,7 +76,7 @@ def signup():
         username = request.form["user"].encode('utf-8')
         if algo.get_user(username) == None:
             salt = bcrypt.gensalt().encode('utf-8')
-            combopass = (request.form["pass"] + salt + master_secret_key).encode('utf-8')
+            combopass = (request.form["pass"] + salt).encode('utf-8')
             password = bcrypt.hashpw(combopass, salt)
             algo.dbWrite_user(username, password, salt)
             message = "New user added"
@@ -159,12 +104,19 @@ def home():
 @app.route("/data", methods = ["POST"])
 def trackdata():
     data = request.get_json()
-    lat = data['lat']
-    lon = data['lon']
+    try:
+        lat = data['lat']
+        lon = data['lon']
+        bike = data['id']
+        if data['time']:
+            time = data['time']
+        else:
+            time = datetime.datetime.now()
+    except TypeError:
+        return("406: incorrect format, accepts JSON for variables 'lat', 'lon', and 'id'")
     if lat and lon:
-        if lat != 0 and lon != 0:
-            print(lat, lon, request.remote_addr)
-            return("200 Success!")
+        print(lat, lon, request.remote_addr)
+        return("200 Success!")
     else:
         return("302 Invalid Request")
 
@@ -177,20 +129,33 @@ def userdata():
         loc_list = [[i.lat, i.lon] for i in l]
         return jsonify(loc_list)
 
-    else:
+    elif request.method == "POST":
         data = request.get_json()
-        lat = data['lat']
-        lon = data['lon']
-        ip = request.remote_addr
-        print(lat, lon, ip)
-            # print(request.form["lat"] + request.form["lon"])
-        last_loc = Location.query.filter(Location.user_ip == ip).first()
-        if last_loc != None:
-            db.session.delete(last_loc)
-        db.session.add(Location(lat, lon, datetime.datetime.utcnow(), ip))
-        db.session.commit()
+        try:
+            lat = data['lat']
+            lon = data['lon']
+            if data['time']:
+                time = data['time']
+            else:
+                time = datetime.datetime.now()
+        except TypeError:
+            return("406: incorrect format, accepts JSON for variables 'lat', 'lon', and 'time'")
+        if session['user']:
+            user = session['user']
 
-        print(Location.query.order_by('-id').first().lat)
+            # log user data
+
+
+            ip = request.remote_addr
+            print(lat, lon, ip)
+                # print(request.form["lat"] + request.form["lon"])
+            last_loc = Location.query.filter(Location.user_ip == ip).first()
+            if last_loc != None:
+                db.session.delete(last_loc)
+            db.session.add(Location(lat, lon, datetime.datetime.utcnow(), ip))
+            db.session.commit()
+
+            print(Location.query.order_by('-id').first().lat)
 
         """
         ips = [u.ip for u in User.query.all()]
